console-window component styling and functionality adjustments
The console-window component now can take in 2 inputs: a string for the Title and a string for the window's innerHTML.

The second value is treated as innerHTML so that document markup (such as <p> tags) can be utilized inside the window section of the component.

In order for stylings to be applied to that innerHTML, the console-window component now utilizes a ViewEncapsulation of ShadowDom. This allows for HTML generated outside the component itself to be styled according to the component's stylings.

Additionally, the string value passed into the component for the window's innerHTML is sanitized using DomSanitizer.