Began adding functionality for non-console command text inside the console-window component
Started playing around with using the console-window in the context of other elements.

Started the process of creating informative non-command blocks of text that can be placed within console windows (to explain the command functionality in detail or in the future to include details such as command options).

Also adjusted the icon elements for the console-window component to not be viewable to screenreaders and not affect the mouse cursor since they are purely decorative.