Created ConsoleWindow helper class
ConsoleWindow helper class now exists in the console-window-component.ts file. This class helps handle the creation of the window text by allowing new lines of text to be added like a stack via a push method. The push method also allows you to specify if the new line is a "Command" or an "Informative" line (css file updated to account for these changes).