Color helper class can take hex code values
Color helper class has been upgraded to now be able to take Hex Codes as an input to the constructor. This means the constructor can take RGB, RGBA, Hex Code with Alpha, and Hex Code without Alpha. It can even interpret short hand hex codes (like #0f0 for #00ff00 or #3A8F for #33AA88FF).

There is also an exception class custom to the color class: the HexCodeError which is thrown when the developer tries to provide invalid hex code values to the constructor.