Small changes to console-window and wordart components
The console-window component has had some small adjustments to its styling. The background color of the command sections now belongs to the entire command window. The titlebar and informative sections both have background-colors that are a transparent white which modifies the main console-window's bg-color.

The elements comprising the wordart component have been modified so that the first layer of text will be placed inside the DOM flow.